// Nitro/src/client/room/object/visualization/IRoomObjectSprite.ts
export interface IRoomObjectSprite {
    visible: boolean;
    relativeDepth: number;
    posture: string;
    height: number;
    width: number;
    direction: number;
    type: any;
    spriteType: any;
    name: string;
    flipV: any;
    offsetY: number;
    offsetX: number;
    flipH: any;
    texture: PIXI.Texture;
    filters: PIXI.Filter[];
    blendMode: number;
    color: number;
    alpha: number;
    clickHandling: boolean;
    alphaTolerance: number;
    tag: string;
    id: number;
    updateCounter: number;
    _Str_4593: boolean;
    _Str_3582: string;

}